feat(wasm): web worker + frontend rewire to in-browser WASM compute
static/worker.js: ES module worker running goc_wasm (initSync+fetch, onmessage
set before init with a pre-ready queue to avoid the module-worker top-level-
await message-loss race). Handles config/optimize/distribution, returning
the exact /api JSON shapes so Chart.js rendering is unchanged.
templates/index.html: <script type=module>, creates the worker, replaces the
3 fetch('/api/...') calls with wasmCall (postMessage RPC). Compute now runs
off the main thread in WASM.
stats.rs/wasm binding: pity99 -> f64 (Infinity sentinel) so the worker maps
inf -> null via isFinite, matching Python float('inf') -> null.

E2E (Playwright, Flask serving /static): config loads 46 events/5 formulas;
optimize (f_count=5,target=10,n_max=30) -> best_n=21, prob=0.13835902
(matches planner golden); curve-click -> distribution at n=13. 0 console
errors. main.py CLI untouched.

+++ bindings/wasm/src/lib.rs
@@ -67,6 +67,6 @@ pub fn expected_attempts(p: f64) -> f64 {
 }
 
 #[wasm_bindgen]
-pub fn pity99(p: f64) -> usize {
+pub fn pity99(p: f64) -> f64 {
     stats::pity99(p)
 }
